backend/API/requirements.txt
psycopg[binary]
python-dotenv
fastapi
uvicorn
httpx